[ALL][ISSUE-50]: second part of refactoring
