+++ akka-http/src/test/scala/com/github/fsanaulla/chronicler/akka/handlers/AkkaJsonHandlerSpec.scala
@@ -2,20 +2,22 @@ package com.github.fsanaulla.chronicler.akka.handlers
 
 import _root_.akka.actor.ActorSystem
 import _root_.akka.http.scaladsl.model.{HttpEntity, HttpResponse}
-import _root_.akka.stream.ActorMaterializer
-import com.github.fsanaulla.chronicler.akka.utils.AkkaContentTypes._
+import akka.testing.Testkit
+import com.github.fsanaulla.chronicler.akka.utils.AkkaContentTypes.AppJson
 import com.github.fsanaulla.core.test.TestSpec
-import jawn.ast.JParser
+import jawn.ast._
 
-import scala.concurrent.ExecutionContext
+class AkkaJsonHandlerSpec
+  extends TestKit(ActorSystem())
+    with TestSpec
+    with AkkaJsonHandler {
 
-class AkkaJsonHandlerSpec extends TestSpec with AkkaJsonHandler {
+  override def afterAll: Unit = {
+    TestKit.shutdownActorSystem(system)
+  }
 
-  implicit val system: ActorSystem = ActorSystem("TestOne")
-  implicit val mat: ActorMaterializer = ActorMaterializer()
-  implicit val ex: ExecutionContext = system.dispatcher
-
-  val singleStrJson = """{
+  "AkkaJsonHandler" should "extract js object from HTTP response" in {
+    val singleStrJson = """{
                       "results": [
                           {
                               "statement_id": 0,
@@ -46,11 +48,184 @@ class AkkaJsonHandlerSpec extends TestSpec with AkkaJsonHandler {
                       ]
                   }"""
 
-  val singleHttpResponse: HttpResponse = HttpResponse(entity = HttpEntity(AppJson, singleStrJson))
+  val singleHttpResponse: HttpResponse =
+    HttpResponse(entity = HttpEntity(AppJson, singleStrJson))
 
-  val jsResult = JParser.parseFromString(singleStrJson).toOption.value
+  val jsResult: JValue =
+    JParser.parseFromString(singleStrJson).toOption.value
 
-  "Akka json handler" should "extract js object from HTTP response" in {
-    getJsBody(singleHttpResponse).futureValue shouldEqual jsResult
+    getResponseBody(singleHttpResponse).futureValue shouldEqual jsResult
+  }
+
+  it should "extract single query result from JSON" in {
+
+    val json  = JParser.parseFromString(
+      """
+        |{
+        |    "results": [
+        |        {
+        |            "statement_id": 0,
+        |            "series": [
+        |                {
+        |                    "name": "cpu_load_short",
+        |                    "columns": [
+        |                        "time",
+        |                        "name",
+        |                        "value"
+        |                    ],
+        |                    "values": [
+        |                        [
+        |                            "2015-01-29T21:55:43.702900257Z",
+        |                            "Fz",
+        |                            2
+        |                        ],
+        |                        [
+        |                            "2015-01-29T21:55:43.702900257Z",
+        |                            "Rz",
+        |                            0.55
+        |                        ],
+        |                        [
+        |                            "2015-06-11T20:46:02Z",
+        |                            null,
+        |                            0.64
+        |                        ]
+        |                    ]
+        |                }
+        |            ]
+        |        }
+        |    ]
+        |}
+      """.stripMargin).toOption.value
+
+    val result = Array(
+      JArray(Array(JString("2015-01-29T21:55:43.702900257Z"), JString("Fz"), JNum(2))),
+      JArray(Array(JString("2015-01-29T21:55:43.702900257Z"), JString("Rz"), JNum(0.55))),
+      JArray(Array(JString("2015-06-11T20:46:02Z"), JNull, JNum(0.64)))
+    )
+
+    getOptInfluxPoints(json).value shouldEqual result
+  }
+
+  it should "extract bulk query result from JSON" in {
+    val json = JParser.parseFromString(
+      """
+        |{
+        |    "results": [
+        |        {
+        |            "statement_id": 0,
+        |            "series": [
+        |                {
+        |                    "name": "cpu_load_short",
+        |                    "columns": [
+        |                        "time",
+        |                        "value"
+        |                    ],
+        |                    "values": [
+        |                        [
+        |                            "2015-01-29T21:55:43.702900257Z",
+        |                            2
+        |                        ],
+        |                        [
+        |                            "2015-01-29T21:55:43.702900257Z",
+        |                            0.55
+        |                        ],
+        |                        [
+        |                            "2015-06-11T20:46:02Z",
+        |                            0.64
+        |                        ]
+        |                    ]
+        |                }
+        |            ]
+        |        },
+        |        {
+        |            "statement_id": 1,
+        |            "series": [
+        |                {
+        |                    "name": "cpu_load_short",
+        |                    "columns": [
+        |                        "time",
+        |                        "count"
+        |                    ],
+        |                    "values": [
+        |                        [
+        |                            "1970-01-01T00:00:00Z",
+        |                            3
+        |                        ]
+        |                    ]
+        |                }
+        |            ]
+        |        }
+        |    ]
+        |}
+      """.stripMargin).toOption.value
+
+    val result = Array(
+      Array(
+        JArray(Array(JString("2015-01-29T21:55:43.702900257Z"), JNum(2))),
+        JArray(Array(JString("2015-01-29T21:55:43.702900257Z"), JNum(0.55))),
+        JArray(Array(JString("2015-06-11T20:46:02Z"), JNum(0.64)))),
+      Array(
+        JArray(Array(JString("1970-01-01T00:00:00Z"), JNum(3)))
+      )
+    )
+
+    getOptBulkInfluxPoints(json).value shouldEqual result
+  }
+
+  it should "extract influx information from JSON" in {
+    val json  = JParser.parseFromString(
+      """
+        |{
+        |    "results": [
+        |        {
+        |            "statement_id": 0,
+        |            "series": [
+        |                {
+        |                    "name": "cpu_load_short",
+        |                    "columns": [
+        |                        "time",
+        |                        "value"
+        |                    ],
+        |                    "values": [
+        |                        [
+        |                            "2015-01-29T21:55:43.702900257Z",
+        |                            2
+        |                        ],
+        |                        [
+        |                            "2015-01-29T21:55:43.702900257Z",
+        |                            0.55
+        |                        ],
+        |                        [
+        |                            "2015-06-11T20:46:02Z",
+        |                            0.64
+        |                        ]
+        |                    ]
+        |                }
+        |            ]
+        |        }
+        |    ]
+        |}
+      """.stripMargin).toOption.value
+
+    val result = Array(
+      "cpu_load_short" -> Array(
+        JArray(Array(JString("2015-01-29T21:55:43.702900257Z"), JNum(2))),
+        JArray(Array(JString("2015-01-29T21:55:43.702900257Z"), JNum(0.55))),
+        JArray(Array(JString("2015-06-11T20:46:02Z"), JNum(0.64)))
+      )
+    )
+
+    val res = getOptJsInfluxInfo(json)
+
+    res should not be None
+    res.value.length shouldEqual 1
+    val (measurament, points) = res.value.head
+
+    measurament shouldEqual "cpu_load_short"
+    points shouldEqual Array(
+      JArray(Array(JString("2015-01-29T21:55:43.702900257Z"), JNum(2))),
+      JArray(Array(JString("2015-01-29T21:55:43.702900257Z"), JNum(0.55))),
+      JArray(Array(JString("2015-06-11T20:46:02Z"), JNum(0.64)))
+    )
   }
 }
