+++ core/src/main/scala/com/github/fsanaulla/core/api/DatabaseApi.scala
@@ -15,10 +15,10 @@ private[fsanaulla] abstract class DatabaseApi[M[_], E](dbName: String)
   extends ReadOperations[M] with WriteOperations[M, E] {
 
   def read[A: ClassTag](query: String,
-                              epoch: Epoch = Epochs.NANOSECONDS,
-                              pretty: Boolean = false,
-                              chunked: Boolean = false)
-                             (implicit reader: InfluxReader[A]): M[QueryResult[A]]
+                        epoch: Epoch = Epochs.NANOSECONDS,
+                        pretty: Boolean = false,
+                        chunked: Boolean = false)
+                       (implicit reader: InfluxReader[A]): M[QueryResult[A]]
 
   final def writeFromFile0(file: File,
                            chunkSize: Int = 8192,
@@ -26,7 +26,7 @@ private[fsanaulla] abstract class DatabaseApi[M[_], E](dbName: String)
                            precision: Precision = Precisions.NANOSECONDS,
                            retentionPolicy: Option[String] = None)
                           (implicit ds: Deserializer[File, E]): M[Result] = {
-    write0(dbName, ds.deserialize(file), consistency, precision, retentionPolicy)
+    writeTo(dbName, ds.deserialize(file), consistency, precision, retentionPolicy)
   }
 
   final def writeNative0(point: String,
@@ -34,7 +34,7 @@ private[fsanaulla] abstract class DatabaseApi[M[_], E](dbName: String)
                          precision: Precision = Precisions.NANOSECONDS,
                          retentionPolicy: Option[String] = None)
                         (implicit ds: Deserializer[String, E]): M[Result] = {
-    write0(dbName, ds.deserialize(point), consistency, precision, retentionPolicy)
+    writeTo(dbName, ds.deserialize(point), consistency, precision, retentionPolicy)
   }
 
   final def bulkWriteNative0(points: Seq[String],
@@ -42,7 +42,7 @@ private[fsanaulla] abstract class DatabaseApi[M[_], E](dbName: String)
                              precision: Precision = Precisions.NANOSECONDS,
                              retentionPolicy: Option[String] = None)
                             (implicit ds: Deserializer[Seq[String], E]): M[Result] = {
-    write0(dbName, ds.deserialize(points), consistency, precision, retentionPolicy)
+    writeTo(dbName, ds.deserialize(points), consistency, precision, retentionPolicy)
   }
 
   final def writePoint0(point: Point,
@@ -50,7 +50,7 @@ private[fsanaulla] abstract class DatabaseApi[M[_], E](dbName: String)
                         precision: Precision = Precisions.NANOSECONDS,
                         retentionPolicy: Option[String] = None)
                        (implicit ds: Deserializer[Point, E]): M[Result] = {
-    write0(dbName, ds.deserialize(point), consistency, precision, retentionPolicy)
+    writeTo(dbName, ds.deserialize(point), consistency, precision, retentionPolicy)
   }
 
   final def bulkWritePoints0(points: Seq[Point],
@@ -58,7 +58,7 @@ private[fsanaulla] abstract class DatabaseApi[M[_], E](dbName: String)
                              precision: Precision = Precisions.NANOSECONDS,
                              retentionPolicy: Option[String] = None)
                             (implicit ds: Deserializer[Seq[Point], E]): M[Result] = {
-    write0(dbName, ds.deserialize(points), consistency, precision, retentionPolicy)
+    writeTo(dbName, ds.deserialize(points), consistency, precision, retentionPolicy)
   }
 
   final def readJs(query: String,
