+++ project/Dependencies.scala
@@ -9,6 +9,7 @@ object Dependencies {
 
   object Versions {
     final val sttp = "1.1.14"
+    final val akka = "2.5.12"
   }
 
   // core
@@ -21,8 +22,10 @@ object Dependencies {
 
   // akka-http
   final val akkaDep = Seq(
-    "com.typesafe.akka" %% "akka-stream" % "2.5.11" % Provided,
-    "com.typesafe.akka" %% "akka-http"   % "10.1.1"
+    "com.typesafe.akka" %% "akka-stream"  % Versions.akka % Provided,
+    "com.typesafe.akka" %% "akka-actor"   % Versions.akka % Provided,
+    "com.typesafe.akka" %% "akka-testkit" % Versions.akka % Test,
+    "com.typesafe.akka" %% "akka-http"    % "10.1.1"
   )
 
   // async-http
